chore: refactor SearchSettingsDropdown more
Removing unused props simplifies the logic a lot.
