+++ app/component/itinerary/customizesearch/BikingSpeed.js
@@ -1,16 +1,23 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { saveRoutingSettings } from '../../../action/SearchSettingsActions';
-import SearchSettingsDropdown, {
-  getFiveStepOptionsNumerical,
-  valueShape,
-} from './SearchSettingsDropdown';
+import SearchSettingsDropdown, { valueShape } from './SearchSettingsDropdown';
 import { addAnalyticsEvent } from '../../../util/analyticsUtils';
 import { findNearestOption } from '../../../util/planParamUtil';
-import { settingsShape } from '../../../util/shapes';
+
+const getFiveStepOptionsNumerical = options => {
+  const numericalOptions = [];
+  options.forEach(item => {
+    numericalOptions.push({
+      title: `${Math.round(item * 3.6)} km/h`,
+      value: item,
+    });
+  });
+  return numericalOptions;
+};
 
 export default function BikingSpeed(
-  { bikeSpeed, bikeSpeedOptions, defaultSettings },
+  { bikeSpeed, bikeSpeedOptions },
   { executeAction },
 ) {
   const options = getFiveStepOptionsNumerical(bikeSpeedOptions);
@@ -23,7 +30,6 @@ export default function BikingSpeed(
     <SearchSettingsDropdown
       name="bike-speed-selector"
       currentSelection={currentSelection}
-      defaultValue={defaultSettings.bikeSpeed}
       onOptionSelected={value => {
         executeAction(saveRoutingSettings, {
           bikeSpeed: value,
@@ -35,7 +41,6 @@ export default function BikingSpeed(
         });
       }}
       options={options}
-      formatOptions
       labelId="biking-speed"
       translateLabels={false}
     />
@@ -45,7 +50,6 @@ export default function BikingSpeed(
 BikingSpeed.propTypes = {
   bikeSpeed: valueShape.isRequired,
   bikeSpeedOptions: PropTypes.arrayOf(PropTypes.number).isRequired,
-  defaultSettings: settingsShape.isRequired,
 };
 
 BikingSpeed.contextTypes = {
