+++ app/component/itinerary/customizesearch/SearchSettingsDropdown.js
@@ -4,57 +4,6 @@ import { FormattedMessage } from 'react-intl';
 import Icon from '../../Icon';
 import { useTranslationsContext } from '../../../util/useTranslationsContext';
 
-const roundToOneDecimal = number => {
-  const rounded = Math.round(number * 10) / 10;
-  return rounded.toFixed(1).replace('.', ',');
-};
-
-/**
- * Builds an array of options: least, less, default, more, most with preset
- * multipliers or given values for each option. Note: a higher value (relative to
- * the given value) means less/worse.
- *
- * @param {*} options The options to select from.
- */
-export const getFiveStepOptions = options => [
-  {
-    title: 'option-least',
-    value: options.least || options[0],
-    kmhValue: `${roundToOneDecimal(options[0] * 3.6)} km/h`,
-  },
-  {
-    title: 'option-less',
-    value: options.less || options[1],
-    kmhValue: `${roundToOneDecimal(options[1] * 3.6)} km/h`,
-  },
-  {
-    title: 'option-default',
-    value: options[2],
-    kmhValue: `${roundToOneDecimal(options[2] * 3.6)} km/h`,
-  },
-  {
-    title: 'option-more',
-    value: options.more || options[3],
-    kmhValue: `${roundToOneDecimal(options[3] * 3.6)} km/h`,
-  },
-  {
-    title: 'option-most',
-    value: options.most || options[4],
-    kmhValue: `${roundToOneDecimal(options[4] * 3.6)} km/h`,
-  },
-];
-
-export const getFiveStepOptionsNumerical = options => {
-  const numericalOptions = [];
-  options.forEach(item => {
-    numericalOptions.push({
-      title: `${Math.round(item * 3.6)} km/h`,
-      value: item,
-    });
-  });
-  return numericalOptions;
-};
-
 /**
  * Represents the types of acceptable values.
  */
@@ -68,13 +17,8 @@ export const valueShape = PropTypes.oneOfType([
 export default function SearchSettingsDropdown({
   labelId,
   options,
-  displayValueFormatter,
   currentSelection,
-  highlightDefaultValue,
-  defaultValue,
-  displayPattern,
   onOptionSelected,
-  formatOptions,
   name,
   translateLabels,
 }) {
@@ -146,56 +90,16 @@ export default function SearchSettingsDropdown({
     ));
   };
 
-  const applyDefaultValueIdentifier = (value, str) =>
-    highlightDefaultValue && value === defaultValue
-      ? `${intl.formatMessage({
-          id: 'option-default',
-        })} (${str})`
-      : str;
-
-  const getFormattedValue = value =>
-    displayValueFormatter ? displayValueFormatter(value) : value;
-
-  const selectOptions = formatOptions
-    ? options.map(o =>
-        o.title && o.value
-          ? {
-              displayName: `${o.title}_${o.value}`,
-              displayNameObject: applyDefaultValueIdentifier(
-                o.value,
-                translateLabels
-                  ? intl.formatMessage(
-                      { id: o.title },
-                      {
-                        title: o.title,
-                      },
-                    )
-                  : o.title,
-              ),
-              value: o.value,
-              kmhValue: o.kmhValue || undefined,
-            }
-          : {
-              displayName: `${displayPattern}_${o}`,
-              displayNameObject: applyDefaultValueIdentifier(
-                o,
-                // eslint-disable-next-line no-nested-ternary
-                displayPattern
-                  ? translateLabels
-                    ? intl.formatMessage(
-                        { id: displayPattern },
-                        {
-                          number: getFormattedValue(o),
-                        },
-                      )
-                    : ({ id: displayPattern }, { number: getFormattedValue(o) })
-                  : getFormattedValue(o),
-              ),
-              value: o,
-              kmhValue: o.kmhValue || undefined,
-            },
-      )
-    : options;
+  const selectOptions = options.map(o => {
+    return {
+      displayName: `${o.title}_${o.value}`,
+      displayNameObject: translateLabels
+        ? intl.formatMessage({ id: o.title }, { title: o.title })
+        : o.title,
+      value: o.value,
+      kmhValue: o.kmhValue,
+    };
+  });
 
   return (
     <div className="settings-dropdown-wrapper" ref={labelRef}>
@@ -207,14 +111,11 @@ export default function SearchSettingsDropdown({
         <FormattedMessage id={labelId} />
         <span className="settings-dropdown-text-container">
           <p className="settings-dropdown-label-value">
-            {/* eslint-disable-next-line no-nested-ternary */}
-            {displayValueFormatter
-              ? displayValueFormatter(currentSelection.title)
-              : translateLabels
-                ? `${intl.formatMessage({
-                    id: currentSelection.title,
-                  })}`
-                : currentSelection.title}
+            {translateLabels
+              ? `${intl.formatMessage({
+                  id: currentSelection.title,
+                })}`
+              : currentSelection.title}
           </p>
           <span
             aria-label={intl.formatMessage({
@@ -243,25 +144,15 @@ export default function SearchSettingsDropdown({
 SearchSettingsDropdown.propTypes = {
   labelId: PropTypes.string.isRequired,
   options: PropTypes.arrayOf(valueShape).isRequired,
-  displayValueFormatter: PropTypes.func,
   currentSelection: PropTypes.shape({
     title: PropTypes.string,
     value: valueShape,
   }).isRequired,
-  highlightDefaultValue: PropTypes.bool,
-  defaultValue: valueShape,
-  displayPattern: PropTypes.string,
   onOptionSelected: PropTypes.func.isRequired,
-  formatOptions: PropTypes.bool,
   name: PropTypes.string.isRequired,
   translateLabels: PropTypes.bool,
 };
 
 SearchSettingsDropdown.defaultProps = {
-  displayValueFormatter: undefined,
-  highlightDefaultValue: false,
-  displayPattern: undefined,
-  defaultValue: undefined,
-  formatOptions: false,
   translateLabels: true,
 };
