+++ app/component/itinerary/customizesearch/WalkingOptions.js
@@ -2,13 +2,51 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { saveRoutingSettings } from '../../../action/SearchSettingsActions';
 import { addAnalyticsEvent } from '../../../util/analyticsUtils';
-import SearchSettingsDropdown, {
-  getFiveStepOptions,
-} from './SearchSettingsDropdown';
+import SearchSettingsDropdown from './SearchSettingsDropdown';
 import SettingsToggle from './SettingsToggle';
 import { findNearestOption } from '../../../util/planParamUtil';
 import { settingsShape } from '../../../util/shapes';
 import { useConfigContext } from '../../../configurations/ConfigContext';
+
+const roundToOneDecimal = number => {
+  const rounded = Math.round(number * 10) / 10;
+  return rounded.toFixed(1).replace('.', ',');
+};
+
+/**
+ * Builds an array of options: least, less, default, more, most with preset
+ * multipliers or given values for each option. Note: a higher value (relative to
+ * the given value) means less/worse.
+ *
+ * @param {*} options The options to select from.
+ */
+const getFiveStepOptions = options => [
+  {
+    title: 'option-least',
+    value: options.least || options[0],
+    kmhValue: `${roundToOneDecimal(options[0] * 3.6)} km/h`,
+  },
+  {
+    title: 'option-less',
+    value: options.less || options[1],
+    kmhValue: `${roundToOneDecimal(options[1] * 3.6)} km/h`,
+  },
+  {
+    title: 'option-default',
+    value: options[2],
+    kmhValue: `${roundToOneDecimal(options[2] * 3.6)} km/h`,
+  },
+  {
+    title: 'option-more',
+    value: options.more || options[3],
+    kmhValue: `${roundToOneDecimal(options[3] * 3.6)} km/h`,
+  },
+  {
+    title: 'option-most',
+    value: options.most || options[4],
+    kmhValue: `${roundToOneDecimal(options[4] * 3.6)} km/h`,
+  },
+];
 
 export default function WalkingOptions(
   { currentSettings, defaultSettings },
@@ -42,7 +80,6 @@ export default function WalkingOptions(
     <>
       <SearchSettingsDropdown
         currentSelection={currentWalkSelection}
-        defaultValue={defaultSettings.walkSpeed}
         onOptionSelected={value => {
           executeAction(saveRoutingSettings, {
             walkSpeed: value,
@@ -55,8 +92,6 @@ export default function WalkingOptions(
         }}
         options={options}
         labelId="walking-speed"
-        highlightDefaulValue
-        formatOptions
         name="walkspeed"
       />
       <SettingsToggle
